Move context validators
